Armory, gun and item overhaul
Changes how items/guns are handled by the system and added the armory to equip different guns that are unlocked through play. Fixed the weapon display UI to update on gun change and accommodate a single gun. Shop now uses scriptable objects with reference to a shop model and actual item prefab.